Move everything into its own package
