+++ cmd/soroban-rpc/internal/xdr2json/conversion.go
@@ -1,11 +1,11 @@
-package preflight
+package xdr2json
 
 /*
 // See preflight.go for explanations:
 
-#include "../../lib/preflight.h"
-#include "../../lib/xdrjson.h"
 #include <stdlib.h>
+#include "../../lib/xdrjson.h"
+
 #cgo windows,amd64 LDFLAGS: -L${SRCDIR}/../../../../target/x86_64-pc-windows-gnu/release-with-panic-unwind/ -lpreflight -lntdll -static -lws2_32 -lbcrypt -luserenv
 #cgo darwin,amd64 LDFLAGS: -L${SRCDIR}/../../../../target/x86_64-apple-darwin/release-with-panic-unwind/ -lpreflight -ldl -lm
 #cgo darwin,arm64 LDFLAGS: -L${SRCDIR}/../../../../target/aarch64-apple-darwin/release-with-panic-unwind/ -lpreflight -ldl -lm
@@ -24,9 +24,9 @@ import (
 	"github.com/pkg/errors"
 )
 
-func XdrToJson(xdr interface{}, field []byte) (map[string]interface{}, error) {
+func Convert(xdr interface{}, field []byte) (map[string]interface{}, error) {
 	xdrTypeName := reflect.TypeOf(xdr).Name()
-	goStr := XdrToString(xdrTypeName, field)
+	goStr := ConvertStr(xdrTypeName, field)
 
 	var result map[string]interface{}
 	err := json.Unmarshal([]byte(goStr), &result)
@@ -37,7 +37,7 @@ func XdrToJson(xdr interface{}, field []byte) (map[string]interface{}, error) {
 	return result, nil
 }
 
-func AnyXdrToString(xdr interface{}) (string, error) {
+func AnyConvertStr(xdr interface{}) (string, error) {
 	xdrTypeName := reflect.TypeOf(xdr).Name()
 	if cerealXdr, ok := xdr.(encoding.BinaryMarshaler); !ok {
 		data, err := cerealXdr.MarshalBinary()
@@ -45,13 +45,13 @@ func AnyXdrToString(xdr interface{}) (string, error) {
 			return "", errors.Wrapf(err, "")
 		}
 
-		return XdrToString(xdrTypeName, data), nil
+		return ConvertStr(xdrTypeName, data), nil
 	}
 
 	return "", fmt.Errorf("expected serializable XDR, got '%s': %+v", xdrTypeName, xdr)
 }
 
-func XdrToString(xdrTypeName string, field []byte) string {
+func ConvertStr(xdrTypeName string, field []byte) string {
 	var goStr string
 	// scope just added to show matching alloc/frees
 	{
@@ -66,4 +66,11 @@ func XdrToString(xdrTypeName string, field []byte) string {
 	}
 
 	return goStr
+}
+
+func CXDR(xdr []byte) C.xdr_t {
+	return C.xdr_t{
+		xdr: (*C.uchar)(C.CBytes(xdr)),
+		len: C.size_t(len(xdr)),
+	}
 }
